chore: RegionResolver 이름 변경
